ping: merge all ping header files into a single one
 * Rename `ping_common.h` to `ping.h`
 * Merge in `ping6_common.h`
 * Merge in `ping6_niquery.h`
 * Merge in bits from source files
